more conversions

+++ include/boost/multi/adaptors/fftw/test/core.cpp
@@ -219,7 +219,7 @@ auto main() -> int {  // NOLINT(readability-function-cognitive-complexity,bugpro
 			multi::array_ref<complex, 3>(data_elements(io), extensions(io)),
 			multi::array_ref<complex, 3>(data_elements(io), extensions(io))
 		);
-		BOOST_TEST( powerin - power(io)/num_elements(io) < 1e-10 );
+		BOOST_TEST( powerin - power(io)/static_cast<double>(num_elements(io)) < 1e-10 );
 	}
 
 	BOOST_AUTO_TEST_CASE(fftw_2D_const_range_ref) {
